+++ custom_components/aqara_m1s_zigbee_router/const.py
@@ -23,6 +23,7 @@ DATA_SOUND_PLAYERS = "sound_players"
 SERVICE_UPLOAD_SOUND = "upload_sound"
 SERVICE_DELETE_SOUND = "delete_sound"
 SERVICE_REFRESH_SOUNDS = "refresh_sounds"
+SERVICE_RESET_MEDIA_GROUP = "reset_media_group"
 
 SOUND_ROOT = "/data/musics"
 MANAGED_SOUND_ROOT = "/data/musics/music-ch"